Download any file type from DocumentCloud
Basically the same as get-pdf.py except it can be used with any file
type (prompts for file extension, but only affects the file name, not
the data written to that file).

This cannot be used to convert file formats, there are already better
tools for that, such as the legendary Pandoc.